+++ public/userscripts/manga-tracker.user.js
@@ -24,8 +24,10 @@
 // @include      /^https:\/\/jaiminisbox\.com\/reader\/read\/.*?\/[a-z]+\/[0-9]+\/[0-9]+(\/.*)?$/
 // @include      /^https:\/\/kobato\.hologfx\.com\/reader\/read\/.*?\/[a-z]+\/[0-9]+\/[0-9]+(\/.*)?$/
 // @include      /^http:\/\/www\.demonicscans\.com\/FoOlSlide\/read\/.*?\/[a-z]+\/[0-9]+\/[0-9]+(\/.*)?$/
-// @updated      2017-01-15
-// @version      1.3.4
+// @include      /^https?:\/\/reader\.deathtollscans\.net\/read\/.*?\/[a-z]+\/[0-9]+\/[0-9]+(\/.*)?$/
+// @include      /^http:\/\/read\.egscans\.com\/[A-Za-z0-9\-_\!,]+(?:\/Chapter_[0-9]+(?:_extra)?\/?)?$/
+// @updated      2017-01-23
+// @version      1.3.5
 // @downloadURL  https://trackr.moe/userscripts/manga-tracker.user.js
 // @updateURL    https://trackr.moe/userscripts/manga-tracker.meta.js
 // @require      https://ajax.googleapis.com/ajax/libs/jquery/3.1.1/jquery.min.js
@@ -1108,6 +1110,32 @@ let sites = {
 		}
 	}),
 
+	'read.egscans.com' : extendSite({
+		setObjVars : function() {
+			let _this = this;
+
+			this.title       = this.segments[1];
+			this.chapter     = this.segments[2] || 'Chapter_001';
+
+			this.title_url   = 'http://read.egscans.com/'+this.title+'/';
+			this.chapter_url = this.title_url+this.chapter+'/';
+
+			let option = $('select[name=chapter] > option');
+			option.each(function(i, e) {
+				$(e).val(_this.title_url + $(e).val() + '/');
+			});
+			this.chapterList        = generateChapterList(option, 'value');
+			this.chapterListCurrent = this.chapter_url;
+
+			$('script:contains("img_url.push(")').html().match(/url\.push\('(.*?')/g).filter(function(value, index, self) {
+				return self.indexOf(value) === index;
+			}).map(function(v) { return v.substr(10, v.length-11)
+			});
+
+			this.page_count = this.viewerCustomImageList.length;
+		},
+	}),
+
 	'reader.seaotterscans.com' : extendSite({
 		setObjVars : function() {
 			this.title       = this.segments[2];
@@ -1272,6 +1300,38 @@ let sites = {
 
 			this.title_url   = this.https+'://www.demonicscans.com/FoOlSlide/series/'+this.title;
 			this.chapter_url = this.https+'://www.demonicscans.com/FoOlSlide/read/'+this.title+'/'+this.chapter;
+
+			this.chapterList        = generateChapterList($('.topbar_left > .tbtitle:eq(2) > ul > li > a').reverseObj(), 'href');
+			//FIXME: The chapterList isn't properly ordered for series that have chapters in and outside volumes. - https://reader.seaotterscans.com/series/sss/
+			this.chapterListCurrent = this.chapter_url+'/';
+
+			// this.viewerChapterName     = $('.selectChapter:first > option:selected').text().trim();
+			this.viewerTitle           = $('.topbar_left > .dropdown_parent > .text a').text();
+			this.viewerCustomImageList = $('#content').find('> script:first').html().match(/(https?:\\\/\\\/[^"]+)/g).filter(function(value, index, self) {
+				return self.indexOf(value) === index;
+			}).map(function(e) {
+				return e.replace(/\\/g, '');
+			});
+			this.page_count = this.viewerCustomImageList.length;
+		},
+		postSetupTopBar : function() {
+			$('.topbar_left > .tbtitle:eq(2)').remove();
+			$('.topbar_right').remove();
+		},
+		preSetupViewer : function(callback) {
+			$('#page').replaceWith($('<div/>', {id: 'viewer'})); //Set base viewer div
+			callback(true, true);
+		}
+	}),
+
+
+	'reader.deathtollscans.net' : extendSite({
+		setObjVars : function() {
+			this.title       = this.segments[2];
+			this.chapter     = this.segments[3] + '/' + this.segments[4] + '/' + this.segments[5] + (this.segments[6] && this.segments[6] !== 'page' ? '/' + this.segments[6] : '');
+
+			this.title_url   = this.https+'://reader.deathtollscans.net/series/'+this.title;
+			this.chapter_url = this.https+'://reader.deathtollscans.net/read/'+this.title+'/'+this.chapter;
 
 			this.chapterList        = generateChapterList($('.topbar_left > .tbtitle:eq(2) > ul > li > a').reverseObj(), 'href');
 			//FIXME: The chapterList isn't properly ordered for series that have chapters in and outside volumes. - https://reader.seaotterscans.com/series/sss/